test(components-js): move CDN_BASE_URL_DYNAMIC tests from unit to smoke to run on prod build and also check for localhost | bh
