+++ packages/components-js/tests/unit/specs/chunks.spec.ts
@@ -1,7 +1,6 @@
 import * as gzipSize from 'gzip-size';
 import * as path from 'path';
 import * as fs from 'fs';
-import * as globby from 'globby';
 import { COMPONENT_CHUNKS_MANIFEST, ComponentChunkName } from '../../../projects/components-wrapper';
 import { colorExternal } from '@porsche-design-system/components-js/utilities/js';
 import { TAG_NAMES } from '@porsche-design-system/shared';
@@ -346,18 +345,4 @@ describe('chunk content', () => {
       }
     );
   });
-});
-
-describe('package content', () => {
-  const componentsJsFilePath = require.resolve('@porsche-design-system/components-js');
-  const componentsJsPackageDir = path.resolve(componentsJsFilePath, '..');
-  const filePaths = globby.sync(`${componentsJsPackageDir}/**/*.js`);
-
-  it.each(filePaths.map((filePath) => [filePath.replace(componentsJsPackageDir, 'dist/components-wrapper'), filePath]))(
-    'should not contain CDN_BASE_URL_DYNAMIC placeholder in bundled js file: %s',
-    (_, filePath) => {
-      const fileContent = fs.readFileSync(path.resolve(componentsJsPackageDir, filePath), 'utf8');
-      expect(fileContent).not.toContain('CDN_BASE_URL_DYNAMIC');
-    }
-  );
 });
